change devicesbar to a datatypebar (toggle datatypes as opposed to devices)
